requirements.txt
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
reportlab==4.0.4